Optimize worker polling
Add a dedicated producer thread that regularly polls tasks/queues/all
endpoint of conductor and stores a copy of current size of all queues locally.
This is just a simple dict {<taskName>: <queueSize>}

Modify polling function to first check local queue size info and only if
there is >0 tasks in the queue, perform actual poll operation.

The combination above reduces calls to conductor-server poll APIs to 0 in case there are
no tasks to be processed. Also reducing selects in the task table of
postgresql. So idle load on conductor server and the DB is significantly
reduced.

Signed-off-by: Maros Marsalek <mmarsalek@frinx.io>